Ensure that the `fieldname` is preserved.
Previously, if the node did not have text, its `fieldname` was not set.  While
this worked in most cases (there previously was no other type of node that had
a `fieldname`), it causes issues with RangeNodes.  The fix is to always set
`fieldname`.  If the node doesn't need a `fieldname`, this does nothing, so
no issues should be caused.